Remove Postgres related methods from PacketWriter (#694)
